requirements.txt
boto3==1.9.91
botocore==1.12.91
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==2.1.5
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
docutils==0.14
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.1.0
openapi-codec==1.3.2
pdfkit==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
reportlab==3.5.13
requests==2.21.0
s3transfer==0.2.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0
